01_Foundation_ToE/2_Laboratorio_Teorico/PlanckDynamics_Sim/requirements.txt
# Core scientific computing
numpy>=1.24.3
matplotlib>=3.7.1
scipy>=1.10.1
sympy>=1.12
pandas>=2.0.3

# Interactive computing
jupyter>=1.0.0
plotly>=5.15.0
notebook>=6.5.4

# Specialized physics libraries
astropy>=5.3.1          # Astronomy and astrophysics
qutip>=4.7.3            # Quantum computing and quantum information
pyscf>=2.3.0            # Quantum chemistry
fenics>=2019.1.0        # Finite element methods (optional)
h5py>=3.9.0             # HDF5 file format support

# Gravitational waves (optional, may require additional setup)
# gwpy>=3.0.0           # Gravitational wave astronomy

# Machine learning for physics
scikit-learn>=1.3.0     # Machine learning algorithms
tensorflow>=2.13.0       # Deep learning (optional)
pytorch>=2.0.0          # Alternative deep learning (optional)

# High-performance computing
numba>=0.57.0           # JIT compilation
dask>=2023.8.0          # Parallel computing
mpi4py>=3.1.4           # MPI for parallel computing

# Data analysis and visualization
seaborn>=0.12.0         # Statistical visualization
plotnine>=0.10.0        # Grammar of graphics
bokeh>=3.2.0            # Interactive visualizations

# Development and testing
pytest>=7.4.0           # Testing framework
black>=23.7.0           # Code formatting
flake8>=6.0.0           # Code linting
mypy>=1.5.0             # Type checking

# Optional: GPU acceleration
# cupy>=12.0.0          # GPU computing with CUDA
# jax>=0.4.0            # JAX for accelerated computing
